// frontend/src/ipc/generated/TagView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TagSummaryView } from "./TagSummaryView";

export type TagView = { id: string, name: string, revision: number, archived: boolean, merged_into: TagSummaryView | null, task_count: number, series_count: number, life_node_count: number, };